+++ src/components/Conversation/Conversation.vue
@@ -1,13 +1,10 @@
 <template>
   <div v-if="data">
-    <q-alert v-if="data.fetchStatus.hasValidationErrors">
-      {{ data.fetchStatus.validationErrors }}
-    </q-alert>
     <q-infinite-scroll
-      :handler="loadMore"
+      :handler="maybeFetchPast"
     >
       <q-list
-        class="bg-white desktop-margin"
+        class="bg-white desktop-margin relative-position"
       >
         <template v-if="hasLoaded">
           <NotificationToggle
@@ -21,18 +18,24 @@
             @submit="$emit('send', { id: data.id, content: arguments[0] })"
             :placeholder="messagePrompt"
             :user="user"
+            :slim="$q.platform.is.mobile"
           />
           <q-alert
             v-if="data.unreadMessageCount > 0"
             color="secondary"
             icon="star"
+            class="k-unread-alert"
           >
-            {{ $tc('CONVERSATION.UNREAD_MESSAGES', data.unreadMessageCount, { count: data.unreadMessageCount }) }}
-            <q-btn
-              no-caps
-              @click="$emit('markAllRead', data.id)"
-              v-t="'CONVERSATION.MARK_READ'"
-            />
+            <div class="row justify-between items-center">
+              <small>{{ $tc('CONVERSATION.UNREAD_MESSAGES', data.unreadMessageCount, { count: data.unreadMessageCount }) }}</small>
+              <q-btn
+                no-caps
+                outline
+                size="sm"
+                @click="$emit('markAllRead', data.id)"
+                v-t="'CONVERSATION.MARK_READ'"
+              />
+            </div>
           </q-alert>
           <ConversationMessage
             v-for="message in data.messages"
@@ -40,18 +43,16 @@
             :message="message"
             @toggleReaction="$emit('toggleReaction', arguments[0])"
             @save="$emit('saveMessage', arguments[0])"
+            @openThread="$emit('openThread', message)"
           />
         </template>
         <div
-          v-if="data.fetchStatus.pending || data.fetchMoreStatus.pending"
+          v-if="data.fetchStatus.pending || data.fetchPastStatus.pending"
           style="width: 100%; text-align: center">
           <q-spinner-dots :size="40"/>
         </div>
       </q-list>
     </q-infinite-scroll>
-    <q-alert v-if="data.fetchMoreStatus.hasValidationErrors">
-      {{ data.fetchMoreStatus.validationErrors }}
-    </q-alert>
   </div>
 </template>
 
@@ -79,7 +80,7 @@ export default {
       type: Object,
       default: null,
     },
-    fetchMore: {
+    fetchPast: {
       type: Function,
       default: null,
     },
@@ -89,13 +90,13 @@ export default {
     },
   },
   methods: {
-    async loadMore (index, done) {
-      if (!this.data || !this.fetchMore || !this.data.canLoadMore) {
+    async maybeFetchPast (index, done) {
+      if (!this.data || !this.fetchPast || !this.data.canFetchPast) {
         await this.$nextTick()
         done()
         return
       }
-      await this.fetchMore(this.data.id)
+      await this.fetchPast(this.data.id)
       done()
     },
     toggleNotifications () {
@@ -126,7 +127,12 @@ export default {
 
 <style scoped lang="stylus">
 .actionButton
-  float right
-  margin-top -25px
-  margin-right 5px
+  z-index 1
+  position absolute
+  top -24px
+  right 6px
+.k-unread-alert >>>
+  .q-alert-content, .q-alert-side
+    padding-top 6px
+    padding-bottom 6px
 </style>