+++ src/components/Conversation/ConversationMessage.vue
@@ -2,57 +2,65 @@
   <q-item
     v-if="!editMode"
     multiline
-    :class="{ isUnread: message.isUnread }"
-    class="conversation-message"
+    :class="{ isUnread: message.isUnread, slim }"
+    class="conversation-message relative-position"
     highlight
   >
+    <q-btn-group
+      flat
+      class="hover-button k-message-controls"
+    >
+      <q-btn
+        v-if="message.isEditable"
+        flat
+        @click="toggleEdit"
+      >
+        <i class="fas fa-pencil-alt" />
+        <q-tooltip v-t="'BUTTON.EDIT'" />
+      </q-btn>
+      <q-btn
+        v-if="!slim"
+        flat
+        @click="$emit('openThread')"
+      >
+        <i class="fas fa-comments" />
+        <q-tooltip v-t="'CONVERSATION.REPLIES'" />
+      </q-btn>
+      <ConversationAddReaction
+        :reacted="currentUserReactions"
+        @toggle="toggleReaction"
+      />
+    </q-btn-group>
     <q-item-side v-if="!slim">
       <ProfilePicture
         :user="message.author"
-        :size="40"
+        :size="$q.platform.is.mobile ? 30 : 40"
+        style="margin-top: 6px"
       />
     </q-item-side>
     <q-item-main>
-      <q-item-tile class="row no-wrap justify-between items-center">
-        <div class="no-wrap">
-          <router-link :to="{ name: 'user', params: { userId: message.author.id } }">
-            <span class="text-bold text-secondary uppercase">{{ message.author.displayName }}</span>
-          </router-link>
-          <span class="message-date">
-            <small class="text-weight-light">
-              <DateAsWords :date="message.createdAt" />
-            </small>
-          </span>
-          <q-icon
-            v-if="message.receivedVia === 'email'"
-            name="far fa-envelope"
-            class="email-icon"
-          >
-            <q-tooltip v-t="'WALL.RECEIVED_VIA_EMAIL'" />
-          </q-icon>
-        </div>
-        <div>
-          <q-btn
-            v-if="message.isEditable"
-            flat
-            class="hover-button reaction-box self-start"
-            style="opacity: .5; width: 41px; margin-right: 0px"
-            @click="toggleEdit"
-          >
-            <i class="fas fa-pencil-alt" />
-            <q-tooltip v-t="'BUTTON.EDIT'" />
-          </q-btn>
-          <ConversationAddReaction
-            class="hover-button reaction-box self-start"
-            :reacted="currentUserReactions"
-            @toggle="toggleReaction"
-          />
-        </div>
+      <q-item-tile
+        class="no-wrap k-message-meta"
+      >
+        <router-link :to="{ name: 'user', params: { userId: message.author.id } }">
+          <span class="k-message-author text-bold text-secondary uppercase">{{ message.author.displayName }}</span>
+        </router-link>
+        <span class="message-date">
+          <small class="text-weight-light">
+            <DateAsWords :date="message.createdAt" />
+          </small>
+        </span>
+        <q-icon
+          v-if="message.receivedVia === 'email'"
+          name="far fa-envelope"
+          class="email-icon"
+        >
+          <q-tooltip v-t="'WALL.RECEIVED_VIA_EMAIL'" />
+        </q-icon>
       </q-item-tile>
-      <Markdown
-        :source="message.content"
-        class="content"
-      />
+      <div class="content">
+        <Markdown :source="message.content" />
+      </div>
       <div
         v-if="message.isEdited"
         style="margin-top: -5px; opacity: .5"
@@ -72,6 +80,26 @@
         @toggle="toggleReaction"
         style="margin-top: 8px; display: block"
       />
+      <q-btn
+        v-if="showReplies"
+        :outline="message.threadMeta.unreadReplyCount < 1"
+        :color="message.threadMeta.unreadReplyCount > 0 ? 'secondary' : null"
+        @click="$emit('openThread')"
+        class="reaction-box k-thread-box"
+        no-caps
+      >
+        <ProfilePicture
+          class="k-profile-picture"
+          v-for="user in message.threadMeta.participants"
+          :key="user.id"
+          :user="user"
+          :is-link="false"
+        />
+        <span
+          class="k-replies-count"
+          v-t="{ path: 'CONVERSATION.REPLIES_COUNT', choice: message.threadMeta.replyCount, args: { count: message.threadMeta.replyCount } }"
+        />
+      </q-btn>
     </q-item-main>
   </q-item>
   <ConversationCompose
@@ -81,6 +109,7 @@
     @leaveEdit="toggleEdit"
     :user="message.author"
     :value="message.content"
+    :slim="slim"
   />
 </template>
 
@@ -88,14 +117,36 @@
 import ProfilePicture from '@/components/ProfilePictures/ProfilePicture'
 import ConversationReactions from '@/components/Conversation/ConversationReactions'
 import ConversationCompose from '@/components/Conversation/ConversationCompose'
-import { QBtn, QItem, QItemSide, QItemMain, QItemTile, QIcon, QTooltip } from 'quasar'
+import {
+  QBtn,
+  QBtnGroup,
+  QItem,
+  QItemSide,
+  QItemMain,
+  QItemTile,
+  QIcon,
+  QTooltip,
+} from 'quasar'
 import DateAsWords from '@/components/General/DateAsWords'
 import Markdown from '@/components/Markdown'
 import ConversationAddReaction from './ConversationAddReaction'
 export default {
   name: 'ConversationMessage',
   components: {
-    ConversationReactions, ConversationAddReaction, ConversationCompose, ProfilePicture, QBtn, QItem, QItemSide, QItemMain, QItemTile, DateAsWords, QIcon, QTooltip, Markdown,
+    ConversationReactions,
+    ConversationAddReaction,
+    ConversationCompose,
+    ProfilePicture,
+    QBtn,
+    QBtnGroup,
+    QItem,
+    QItemSide,
+    QItemMain,
+    QItemTile,
+    DateAsWords,
+    QIcon,
+    QTooltip,
+    Markdown,
   },
   props: {
     message: {
@@ -115,8 +166,7 @@ export default {
   methods: {
     toggleReaction (name) {
       this.$emit('toggleReaction', {
-        conversationId: this.message.conversation,
-        messageId: this.message.id,
+        message: this.message,
         name,
       })
     },
@@ -140,6 +190,9 @@ export default {
     hasReactions () {
       return this.message && this.message.reactions && this.message.reactions.length > 0
     },
+    showReplies () {
+      return this.message.threadMeta && !this.slim
+    },
   },
 }
 </script>
@@ -148,71 +201,63 @@ export default {
 @import '~variables'
 @import './reactionBox'
 
-// same as PickupItem colors
-$lightGreen = #E7FFE0
-$lighterGreen = #F0FFF0
-
-.left
-  margin-right 1em
-.content
-  word-wrap break-word
-.message-date
-  display inline-block
-  margin-left 2px
-.message-reactions
-  float right
-  color gray
-  padding-bottom 0.1em
 .isUnread
   background linear-gradient(to right, $lightGreen, $lighterGreen)
+
+body.mobile .conversation-message
+  &:not(.slim)
+    padding-left 0
+  >>> .q-item-side
+    min-width 0
+  .k-message-meta
+    font-size 80%
+    padding-top 3px
 .conversation-message
   padding-bottom 0
   .hover-button
     visibility hidden
   &:hover .hover-button
     visibility visible
-.q-item-highlight:hover
-  background-color alpha($secondary, .1)
-.email-icon
-  position relative
-  top -1.5px
-  margin-left 2px
-</style>
+  &.q-item-highlight:hover
+    background-color alpha($secondary, .07)
+  .email-icon
+    position relative
+    top -1.5px
+    margin-left 2px
+  .content
+    word-wrap break-word
+  .message-date
+    display inline-block
+    margin-left 2px
+  .k-thread-box
+    min-height 30px
+    max-height 30px
+    box-shadow none
+    .k-profile-picture
+      margin-right 2px
+      vertical-align middle
+    .k-replies-count
+      margin-left 4px
+      font-size 13px
+      font-weight 500
+      padding-right 3px
+  .k-message-controls
+    position absolute
+    background $secondary
+    top -12px
+    right 8px
+    .q-btn
+      color white
+      transition none
+      padding 2px 9px
+      font-size 13px
+body.desktop
+  .conversation-message.slim .k-message-controls
+    top -8px
+    .q-btn
+      min-height 24px
+      font-size 12px
+  .k-message-meta
+    padding-top 4px
 
-<style lang="stylus">
-// needs to be unscoped because it matches inside rendered markdown
-.conversation-message .parsed
-  overflow-wrap break-word
-  p
-    margin-bottom 0.5rem
-  h1
-    font-size 1.6em
-  h2
-    font-size 1.3em
-  h3
-    font-size 1.2em
-  h4
-    font-size 1.1em
-  h5, h6
-    font-size 1em
-  pre
-    font-size 0.8em
-    padding 3px
-    background-color #ededed
-    margin-top 0.5rem
-    margin-bottom 0.5rem
-    overflow-x auto
-    overflow-wrap normal
-    display block
-  p > code
-    font-size 0.8em
-    padding-left 3px
-    padding-right 3px
-    background-color #ededed
-    overflow-x auto
-    overflow-wrap normal
-    display block
-  img:not(.emoji)
-    max-height 150px
-    max-width 300px
 </style>