+++ src/components/Conversation/EmojiButton.vue
@@ -55,7 +55,7 @@ export default {
 <style scoped lang="stylus">
 .q-btn
   padding 1px 3px
-  min-height 0
+  min-height 23px
 .emoji
   >>> img.emoji
     height 1em