+++ src/pages/Group/Pickups.vue
@@ -66,8 +66,6 @@ export default {
 .notice
   .icon
     margin .1em 0 0 0
-    .fa
-      font-size 10vw
   padding 2em 3em
   transform: translateZ(1px) rotate(-3deg);
   h5