+++ src/pages/MainLayout.vue
@@ -62,6 +62,7 @@
         <q-layout-drawer
           v-if="!$q.platform.is.mobile"
           side="right"
+          :width="400"
           :overlay="false"
           :breakpoint="0"
           :value="showSidenavRight"