+++ src/socket.js
@@ -80,7 +80,13 @@ const socket = {
 
 export function receiveMessage ({ topic, payload }) {
   if (topic === 'conversations:message') {
-    store.dispatch('conversations/receiveMessage', convertMessage(camelizeKeys(payload)))
+    const message = convertMessage(camelizeKeys(payload))
+    if (message.thread) {
+      store.dispatch('currentThread/receiveMessage', message)
+    }
+    if (!message.thread || message.thread === message.id) {
+      store.dispatch('conversations/receiveMessage', message)
+    }
   }
   else if (topic === 'conversations:conversation') {
     store.dispatch('conversations/updateConversation', convertConversation(camelizeKeys(payload)))