+++ CHANGELOG.md
@@ -3,7 +3,29 @@ All notable changes to _karrot_ will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+Please document your changes in this format:
+
+```
+- description of change #PR @username
+```
+
 ## [Unreleased]
+### Added
+
+- Replies to wall messages #1065 @nicksellen @tiltec
+- Link to external route planner for store directions #1020 @pogopaule @tiltec
+- Add notifications about new messages in community.foodsaving.world @tiltec
+
+### Changed
+
+- Hide markdown preview in sidebar @tiltec
+
+### Removed
+
+### Fixed
+
+
+## [Release 5] - 2018-07-16
 ### Added
 
 - web browser push notifications @nicksellen
@@ -150,7 +172,8 @@ After months of planning and programming, the first version of ~the foodsaving t
 - **store**: create a store which belongs to a group, set up name, description, address and position of a store
 - **pick-up dates**: create a pick-up date which belongs to a store, set time and max. amount of people who can collect food, user can join/leave a pick-up date
 
-[Unreleased]: https://github.com/yunity/karrot-frontend/compare/v4.0.0...HEAD
+[Unreleased]: https://github.com/yunity/karrot-frontend/compare/v5.0.0...HEAD
+[Release 5]: https://github.com/yunity/karrot-frontend/compare/v4.0.0...v5.0.0
 [Release 4]: https://github.com/yunity/karrot-frontend/compare/v3.0.0...v4.0.0
 [Release 3]: https://github.com/yunity/karrot-frontend/compare/r2...v3.0.0
 [Release 2]: https://github.com/yunity/karrot-frontend/compare/r1...r2